refactored types to have centralized in types directory. removed duplicate types
